Add webpack, split into different files, ES6
Major update to the gulpfile (now files... again), redoing them in ES6
instead of the old kind, you now have to run them via Yarn or NPM and we
are using Webpack for the JS instead of just concatenating it. Also
updated all the packages that needed to be updated.

Fixed the tests accordingly. See #3 for more information.